% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_publication_record.R
\name{create_publication_record}
\alias{create_publication_record}
\title{Create a publication record data frame for easy plotting}
\usage{
create_publication_record(data, scopus_id)
}
\arguments{
\item{data}{dataframe}

\item{scopus_id}{string}
}
\value{
data frame with year (numeric), number of citations (numeric), corresponding
status (logical) and author name (string)
}
\description{
Create a publication record data frame for easy plotting
}
\examples{
create_publication_record(data, "23056907500")

}
