% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_researcher_data.R
\name{get_researcher_data}
\alias{get_researcher_data}
\title{Get Scopus researcher data}
\usage{
get_researcher_data(scopus_id, apiKey)
}
\arguments{
\item{scopus_id}{string}

\item{apiKey}{string}
}
\value{
data frame
}
\description{
Pass the function a valid Scopus researcher ID and a valid Scopus API key and it
returns a data frame of the recorded publications with relevant metadata.
}
\examples{
get_researcher_data(scopus_id = "0123456789", apiKey = "0123abcd0123abcd")
}
