% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subject_areas.R
\name{get_subject_areas}
\alias{get_subject_areas}
\title{Get subject areas from a set of publications}
\usage{
get_subject_areas(data, apiKey)
}
\arguments{
\item{data}{data frame}
}
\value{
data frame with one column of field names obtained from Scopus
and one column of field frequencies
}
\description{
Pass a data frame generated by get_researcher_data() into this function, and it
looks up the academic subject areas of the journals the research was published in.
}
\examples{
get_subject_areas(data)

}
