% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_collaboration_network.R
\name{plot_collaboration_network}
\alias{plot_collaboration_network}
\title{Plot collaboration network}
\usage{
plot_collaboration_network(network)
}
\arguments{
\item{network}{List of nodes and edges of the network}
}
\value{
network graph of publications
}
\description{
Plot collaboration network
}
\examples{
plot_publication_network(network)
}
