% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_publication_profile.R
\name{plot_publication_profile}
\alias{plot_publication_profile}
\title{Plot publication and citation count as a marginal histogram scatter plot}
\usage{
plot_publication_profile(researcher)
}
\arguments{
\item{researcher}{Data frame with all relevant data for plotting}
}
\value{
marginal histogram scatter plot
}
\description{
Plot publication and citation count as a marginal histogram scatter plot
}
\examples{
plot_publication_profile(researcher)
}
