% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_subject_areas.R
\name{plot_subject_areas}
\alias{plot_subject_areas}
\title{Plot the most active research subjects as an ordered bar chart}
\usage{
plot_subject_areas(subject_areas)
}
\arguments{
\item{subject_areas}{}
}
\value{
Bar chart
}
\description{
Plot the most active research subjects as an ordered bar chart
}
\examples{


}
