% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_researcher_data.R
\name{get_researcher_data}
\alias{get_researcher_data}
\title{Get Scopus researcher data}
\usage{
get_researcher_data(scopus_id, apiKey)
}
\arguments{
\item{scopus_id}{string}

\item{apiKey}{string}
}
\value{
list containing publication data frame and collaboration network
nodes and edges
}
\description{
Pass the function a valid Scopus researcher ID and a valid Scopus API key and
it returns a list containing a data frame of the recorded publications with 
relevant metadata.
}
\examples{
get_researcher_data(
scopus_id = "23056907500", 
apiKey = Sys.getenv("SCOPUS_API_KEY")
)
}
